Got emailLog to send emails properly again.  Drastically reduced fetch size b/c DeadlineExceededErrors kept popping up.  Reducing the fetch size didn't seem to help either though, so I ended up purging the entire ChatMsg db.  refreshUsers no longer errors, and the XmppHandler now implements all the same slash commands as the flash app.

+++ tasks/dbpurge.py
@@ -15,8 +15,11 @@ class DbPurgeTask(webapp.RequestHandler):
         q = db.GqlQuery("SELECT * FROM ChatMsg WHERE date < :1", pastDay)
         results = q.fetch(500)
         db.delete(results)
+        q = db.GqlQuery("SELECT * FROM PrivMsg WHERE date < :1", pastDay)
+        results = q.fetch(500)
+        db.delete(results)
         # clear memcache as well
-        memcache.delete("recentChats")
+        memcache.flush_all()
 
         self.response.headers['Content-Type'] = 'text/html'
 
@@ -24,10 +27,19 @@ class DbPurgeTask(webapp.RequestHandler):
 
 class DbPurgeAllTask(webapp.RequestHandler):
     def get(self):
-        results = ChatMsg.all().fetch(500)
-        db.delete(results)
+        while True:
+            results = ChatMsg.all().fetch(100)
+            if len(results) == 0:
+                break
+            db.delete(results)
+        while True:
+            results = PrivMsg.all().fetch(100)
+            if len(results) == 0:
+                break
+            db.delete(results)
+
         # clear memcache as well
-        memcache.delete("recentChats")
+        memcache.flush_all()
 
         self.response.headers['Content-Type'] = 'text/html'
         self.response.out.write("Deleted all messages.")
